Add files via upload
This code was developed to simulate autonomous vehicles using the CARLA framework. Initially, it connects to the local CARLA server and selects blueprint for a pedestrian. Then, it creates a car, which is put into autopilot mode. The code sets up a camera behind the vehicle and reproduces the camera's perspective.

The main loop of the program runs continuously and involves capturing an image from the camera, adding Gaussian noise, storing the image with noise, applying noise to the car's controls, and displaying the image with noise in an OpenCV window. The user can terminate the execution by pressing the 'q' key.

This code provides a framework for developing and testing control algorithms for autonomous vehicles through simulation in CARLA.
In conclusion, it is a code that tries to confuse the main sensor used by the vehicle we created to see how it will behave.